Bring back `final` argument of `process()`, remove `flush()`
Flushing is tricky. We have to pass some audio to `process()`,
and mark that audio as final, to start the flushing. This will produce
unexpected samples in the output, which is acceptable only in realtime mode.
To support multithreaded mode, we would have to essentialy spinlock until
all audio is processed and received, which is also tricky to get right.

Thus, the original "solution" stays.